fix: api instance header변경 및 이름 변경
