+++ src/utils/apiIntsance.ts
@@ -1,15 +1,15 @@
 import axios from "axios";
 
-export const riotServerApi = axios.create({
+export const riotApi = axios.create({
   baseURL: "https://ddragon.leagueoflegends.com",
-  params: {
-    "X-Riot-Token": process.env.RIOT_API_KEY,
+  headers: {
+    "X-Riot-Token": process.env.NEXT_PUBLIC_RIOT_API_KEY,
   },
 });
 
-export const riotClientApi = axios.create({
+export const rotationApi = axios.create({
   baseURL: "https://kr.api.riotgames.com/lol/platform/v3/champion-rotations",
-  params: {
-    "X-Riot-Token": process.env.RIOT_API_KEY,
+  headers: {
+    "X-Riot-Token": process.env.NEXT_PUBLIC_RIOT_API_KEY,
   },
 });
